Separate process specific code from api
