+++ trio_parallel/__init__.py
@@ -4,6 +4,6 @@ from ._version import __version__
 
 from ._impl import (
     to_process_run_sync as run_sync,
-    BrokenWorkerError,
-    current_default_process_limiter,
+    current_default_worker_limiter,
 )
+from _util import BrokenWorkerError
