+++ trio_parallel/_util.py
@@ -1,0 +1,8 @@
+class BrokenWorkerError(RuntimeError):
+    """Raised when a worker process fails or dies unexpectedly.
+
+    This error is not typically encountered in normal use, and indicates a severe
+    failure of either trio-parallel or the code that was executing in the worker.
+    """
+
+    pass
